ptp: create veth before ipam to support dhcp plugin
The ptp plugin doesn't work with the dhcp IPAM plugin because the ptp
plugin creates the veth (that the dhcp plugin tries run a DHCP client on)
after it runs the IPAM plugin to get the IP address.
So create the veth before trying to run the IPAM plugin.

Includes user conf nominated host veth ifname so that the host system
can bridge it.
Also avoid the route hack so the DHCP leased addressing is more normal.